+++ net.py
@@ -156,7 +156,7 @@ class DecodeBlock(nn.Module):
             x = self.conv_1(x)
             x = self.blur(x)
 
-        x = torch.addcmul(x, value=1.0, tensor1=self.noise_weight_1, tensor2=torch.randn([x.shape[0], 1, x.shape[2], x.shape[3]]))
+        x = torch.addcmul(x, value=1.0, tensor1=self.noise_weight_1, tensor2=torch.randn([x.shape[0], 1, x.shape[2], x.shape[3]]).half())
 
         x = x + self.bias_1
 
@@ -168,7 +168,7 @@ class DecodeBlock(nn.Module):
 
         x = self.conv_2(x)
 
-        x = torch.addcmul(x, value=1.0, tensor1=self.noise_weight_2, tensor2=torch.randn([x.shape[0], 1, x.shape[2], x.shape[3]]))
+        x = torch.addcmul(x, value=1.0, tensor1=self.noise_weight_2, tensor2=torch.randn([x.shape[0], 1, x.shape[2], x.shape[3]]).half())
 
         x = x + self.bias_2
 
@@ -257,7 +257,7 @@ class Discriminator(nn.Module):
         x_prev = self.from_rgb[self.layer_count - (lod - 1) - 1](x_prev)
         x_prev = F.leaky_relu(x_prev, 0.2)
 
-        x = torch.lerp(x_prev, x, blend)
+        x = torch.lerp(x_prev.float(), x.float(), blend).half()
 
         for i in range(self.layer_count - (lod - 1) - 1, self.layer_count):
             x = self.encode_block[i](x)
@@ -340,7 +340,7 @@ class Generator(nn.Module):
         needed_resolution = self.layer_to_resolution[lod]
 
         x_prev = F.interpolate(x_prev, size=needed_resolution)
-        x = torch.lerp(x_prev, x, blend)
+        x = torch.lerp(x_prev.float(), x.float(), blend)
 
         return x
